+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/design/InterfaceIsTypeCheckTest.java
@@ -36,7 +36,9 @@ public class InterfaceIsTypeCheckTest
     @Override
     protected String getPath(String filename) throws IOException {
         return super.getPath("checks" + File.separator
-                + "design" + File.separator + filename);
+                + "design" + File.separator
+                + "interfaceistype" + File.separator
+                + filename);
     }
 
     @Test
