+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/design/interfaceistype/InputInterfaceIsType.java
@@ -2,7 +2,7 @@
 // Test case file for checkstyle.
 // Created: 2001
 ////////////////////////////////////////////////////////////////////////////////
-package com.puppycrawl.tools.checkstyle.checks.design;
+package com.puppycrawl.tools.checkstyle.checks.design.interfaceistype;
 
 /**
  * Test input for InterfaceIsTypeCheck
